Separate lib into modules
